Issue19 (#21)
* Other changes

* Modified saveBook function's parameters to just take in a Book object

* Fixed style

* Made the path files relative

* Renamed FileCollector to LocalBookCollector

* Created a resources folder containing one type of every pic format

* Added and modified the JsonHandler unit test

* Created a unit test for the FilterBooks class

* Fixed grammar

* Fixed grammar

* Added a function for testing purposes only -- deletes the temporary Json files but keeps the images for further use

* Changed list.size == 0 to list.isEmpty

* Made the changes you asked for

* Fixed style

* Fixed style

* Fixed style

* Moved DeleteBook and FilterBook's functions into LocalBookCollector then proceeded to delete those classes

* Handled the case where the directory already exists

* Printed out information

* Fixed style

* Forgot to delete unused tests

* Overloaded findBook function

* Expanded the tests
Tried to implement a logger at the beginning

* Added the test resources and fixed style

* Added checkFile test when comparing books and files

* Fixed bug where the user couldn't enter another author with the same last name

* Fixed a few broken tests and hidden bugs when saving  a Book

* Deleted BookCollection test

* Renamed lclCollector

* Reformatted all the tests to Arrange, Act, Assert

* Fixed style
Fixed #15

* All the error messages are now logged

* Tinkered around with logger
Added a log4j properties to store its results in a file as opposed to the console

* Test

* Updated the main files so that they work with URI
All File instances are now just in LocalBookCollector

* Fixed the tests

* Fixed style

* Got rid of some spaces

* Deleted a few more spaces

* Got rid of some spaces

* Made the methods asynchronous

+++ src/main/java/com/azure/app/LocalBookCollector.java
@@ -4,6 +4,8 @@
 package com.azure.app;
 
 import org.apache.commons.io.FilenameUtils;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
@@ -11,6 +13,7 @@ import javax.imageio.ImageIO;
 import java.awt.image.BufferedImage;
 import java.io.File;
 import java.io.IOException;
+import java.net.URI;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
@@ -26,19 +29,20 @@ class LocalBookCollector implements BookCollection {
     private final Set<String> supportedImageFormats;
     private Flux<Book> jsonBooks;
     private List<File> jsonFiles;
+    private String root;
     private final OptionChecker optionChecker = new OptionChecker();
+    private static Logger logger = LoggerFactory.getLogger(JsonHandler.class);
 
-    LocalBookCollector() {
+    LocalBookCollector(String root) {
+        this.root = root;
         supportedImageFormats = Collections.unmodifiableSet(new HashSet<>(Arrays.asList("gif", "png", "jpg")));
-        File directory = new File(Constants.IMAGE_PATH);
-        if (!directory.exists() && !directory.mkdir()) {
-            throw new IllegalStateException("Couldn't create non-existent JSON directory: "
-                + directory.getAbsolutePath());
+        File directory = new File(Paths.get(root, Constants.IMAGE_PATH).toString());
+        if (!directory.exists() && !directory.mkdirs()) {
+            logger.error("Couldn't create non-existent JSON directory: " + directory.getAbsolutePath());
         }
-        File directoryJSON = new File(Constants.JSON_PATH);
-        if (!directoryJSON.exists() && !directoryJSON.mkdir()) {
-            throw new IllegalStateException("Couldn't create non-existent JSON directory: "
-                + directoryJSON.getAbsolutePath());
+        File directoryJSON = new File(Paths.get(root, Constants.JSON_PATH).toString());
+        if (!directoryJSON.exists() && !directoryJSON.mkdirs()) {
+            logger.error("Couldn't create non-existent JSON directory: " + directoryJSON.getAbsolutePath());
         }
         jsonBooks = initializeBooks().cache();
         jsonFiles = retrieveJsonFiles();
@@ -57,10 +61,11 @@ class LocalBookCollector implements BookCollection {
      */
     private Flux<Book> initializeBooks() {
         try {
-            return Flux.fromStream(Files.walk(Paths.get(Constants.JSON_PATH)))
+            return Flux.fromStream(Files.walk(Paths.get(root, Constants.JSON_PATH)))
                 .filter(f -> f.toFile().getName().endsWith(".json"))
                 .map(path -> Constants.SERIALIZER.fromJSONtoBook(new File(path.toString())));
         } catch (IOException e) {
+            logger.error("Error making Flux: ", e);
             return Flux.error(e);
         }
     }
@@ -72,19 +77,22 @@ class LocalBookCollector implements BookCollection {
      * @param author - Author object of the book
      * @param path   - File containing the cover image of the book
      * @return Mono<Boolean> that determines whether the book got saved or not
-     * true - book was successfully saved
+     * true -  book was successfully saved
      * false - book wasn't saved </Boolean>
      */
     @Override
-    public Mono<Boolean> saveBook(String title, Author author, File path) {
-        final Path fullImagePath = Paths.get(Constants.IMAGE_PATH, author.getLastName(), author.getFirstName(), path.getName());
+    public Mono<Boolean> saveBook(String title, Author author, URI path) {
+        File imagePath = new File(path);
+        final Path fullImagePath = Paths.get(root, Constants.IMAGE_PATH, author.getLastName(),
+            author.getFirstName());
         File imageFile = fullImagePath.toFile();
-        if (!imageFile.getParentFile().exists() && !imageFile.mkdirs()) {
-            System.err.println("Could not create directories for: " + fullImagePath.toString());
+        if (!imageFile.exists() && !imageFile.mkdirs()) {
+            logger.error("Couldn't create directories for: " + imageFile.getAbsolutePath());
         }
-        Book book = new Book(title, author, imageFile);
-        if (saveImage(imageFile.getParentFile(), path) && book.checkBook()) {
-            boolean bookSaved = Constants.SERIALIZER.writeJSON(book);
+        URI savedImage = saveImage(imageFile, imagePath);
+        Book book = new Book(title, author, savedImage);
+        if (optionChecker.checkImage(root, savedImage) && book.checkBook(root)) {
+            boolean bookSaved = Constants.SERIALIZER.writeJSON(book, root);
             jsonBooks = initializeBooks().cache();
             jsonFiles = retrieveJsonFiles();
             return Mono.just(bookSaved);
@@ -92,31 +100,48 @@ class LocalBookCollector implements BookCollection {
         return Mono.just(false);
     }
 
-    private boolean saveImage(File directory, File imagePath) {
+    private URI saveImage(File directory, File imagePath) {
         String extension = FilenameUtils.getExtension(imagePath.getName());
         if (!supportedImageFormats.contains(extension)) {
-            return false;
+            return null;
         }
         try {
             BufferedImage bufferedImage = ImageIO.read(imagePath);
-            File image = Paths.get(directory.getPath(), imagePath.getName()).toFile();
-            return ImageIO.write(bufferedImage, extension, image);
+            File image = new File(Paths.get(directory.getPath(), imagePath.getName()).toString());
+            if (ImageIO.write(bufferedImage, extension, image)) {
+                return image.toURI();
+            }
         } catch (IOException ex) {
-            return false;
+            logger.error("Error saving image: ", ex);
+            return null;
         }
+        return null;
     }
 
-
+    /**
+     * Deletes the book and the file based off its information.
+     *
+     * @param bookToCompare - book that will be deleted
+     * @return Mono<Boolean> determines whether or not book was successfully deleted </Boolean>
+     * true - Book was deleted
+     * false - Book wasn't deleted
+     */
     @Override
-    public boolean deleteBook(Book bookToCompare) {
-        boolean delete = jsonFiles.removeIf(x -> optionChecker.checkFile(x, bookToCompare));
+    public Mono<Boolean> deleteBook(Book bookToCompare) {
+        boolean delete = jsonFiles.removeIf(x -> {
+            boolean result = optionChecker.checkFile(x, bookToCompare);
+            if (result) {
+                x.delete();
+            }
+            return result;
+        });
         if (delete) {
-            bookToCompare.getCover().delete();
+            new File(bookToCompare.getCover()).delete();
             deleteEmptyDirectories();
             jsonBooks = initializeBooks().cache();
-            return true;
+            return Mono.just(true);
         }
-        return false;
+        return Mono.just(false);
     }
 
     private List<File> retrieveJsonFiles() {
@@ -124,7 +149,7 @@ class LocalBookCollector implements BookCollection {
             return walk.map(Path::toFile).filter(f -> f.getName().endsWith(".json"))
                 .collect(Collectors.toList());
         } catch (IOException e) {
-            System.err.println("Exception deleting book file.");
+            logger.error("Exception deleting book file.", e);
             return Collections.emptyList();
         }
     }
@@ -147,11 +172,15 @@ class LocalBookCollector implements BookCollection {
      */
     private void clearFiles(File[] files) {
         for (File file : files) {
-            if (file.isDirectory()) {
-                clearFiles(file.listFiles());
-            }
-            if (file.length() == 0 && !file.getAbsolutePath().endsWith(".json")) {
-                file.delete();
+            if (file == null) {
+                return;
+            } else {
+                if (file.isDirectory()) {
+                    clearFiles(file.listFiles());
+                }
+                if (file.length() == 0 && !file.getAbsolutePath().endsWith(".json")) {
+                    file.delete();
+                }
             }
         }
     }
@@ -163,7 +192,7 @@ class LocalBookCollector implements BookCollection {
      * @return - {@link Flux} of {@link Book} with specified title in the collection
      */
     @Override
-    public Flux<Book> findBookTitle(String title) {
+    public Flux<Book> findBook(String title) {
         return jsonBooks.filter(book -> title.contentEquals(book.getTitle()));
     }
 
@@ -174,9 +203,42 @@ class LocalBookCollector implements BookCollection {
      * @return - {@link Flux} of {@link Book} by specified author in the collection
      */
     @Override
-    public Flux<Book> findBookAuthor(Author author) {
+    public Flux<Book> findBook(Author author) {
         return jsonBooks.filter(book -> author.getFirstName().contentEquals(book.getAuthor().getFirstName())
             && book.getAuthor().getLastName().contentEquals(author.getLastName()));
     }
 
+    /**
+     * Determines whether the Flux is empty or not
+     */
+    @Override
+    public Mono<Boolean> hasBooks() {
+        if (jsonBooks.count().block() == null) {
+            return Mono.just(false);
+        }
+        return Mono.just(jsonBooks.count().block() > 0);
+    }
+
+    /**
+     * Determines if an entry is a file
+     */
+    boolean isFile(URI entry) {
+        if (entry == null) {
+            return false;
+        }
+        File fh = new File(entry);
+        return fh.isFile();
+    }
+
+    /**
+     * Converts a string to a File and then returns the file as a uri for the
+     * Book image
+     *
+     * @param path - String containing the image file the user entered
+     * @return URI - created from the converted file
+     */
+    @Override
+    public URI retrieveURI(String path) {
+        return new File(path).toURI();
+    }
 }
